* origin/master: (162 commits)
  BF: monkey patch scipy with broken fit for 2 fixed params (e.g. floc and fscale)
  BF: match_distributino - use "resolved" dist_gen, which might original be a tuple
  BF: workaround for sphinx -- import Process as private __Process from psutil
  BF: fix legend in plot match distributions
  ENH: suppress Python warnings while match_distribution
  ENH/BF: compute unique only if not known + handle all obscure cases via ndarray(list(set()))
  ENH: just an additional externals check to see if np.unique is flawed
  BF: h5py external is required for results_backend="hdf5"
  ENH: run unittest-py3 for origin/master and tiziano/master +store output for examples +safe guard WTF call
  PY3: make SequenceCollectable.unique faster.
  BF: assure different filenames in the SL's _proc_block by providing index of the block (should fix gh 87)
  ENH: make h5py version available
  DOC: Start a changelog for 2.2?
  BF: Remove special case for builtin reconstructors in HDF5 (Closes gh-86)
  Document the HDF5 cannot handle slice()
  DOC: Add reference to Emanuele's conference paper.
  BF: hdf5 -- flatten and save attr 'shape' for object ndarrays (Closes gh-84)
  ENH: fix up the searchlight test for Swaroop's intriguing measure
  Modified searchlight unittest to show hdf5 & native modes differ
  ENH: results_backend to alter how results are passed from pprocess
  ...

Conflicts: primarily due to added in master .ca.roi_feature_ids and heavy
           RF of the test_spatial_searchlight  in _tent/mnn branch
	mvpa2/measures/gnbsearchlight.py
	mvpa2/tests/test_searchlight.py

+++ mvpa2/measures/searchlight.py
@@ -14,11 +14,17 @@ if __debug__:
     from mvpa2.base import debug
 
 import numpy as np
+import tempfile, os
 
 from mvpa2.base import externals, warning
 from mvpa2.base.dochelpers import borrowkwargs, _repr_attrs
+from mvpa2.base.types import is_datasetlike
+if externals.exists('h5py'):
+    # Is optionally required for passing searchlight
+    # results via storing/reloading hdf5 files
+    from mvpa2.base.hdf5 import h5save, h5load
 
-from mvpa2.datasets import hstack
+from mvpa2.datasets import hstack, Dataset
 from mvpa2.support import copy
 from mvpa2.featsel.base import StaticFeatureSelection
 from mvpa2.measures.base import Measure
@@ -35,6 +41,9 @@ class BaseSearchlight(Measure):
 
     roi_sizes = ConditionalAttribute(enabled=False,
         doc="Number of features in each ROI.")
+
+    roi_feature_ids = ConditionalAttribute(enabled=False,
+        doc="Feature IDs for all generated ROIs.")
 
     is_trained = True
     """Indicate that this measure is always trained."""
@@ -61,7 +70,7 @@ class BaseSearchlight(Measure):
       """
         Measure.__init__(self, **kwargs)
 
-        if nproc > 1 and not externals.exists('pprocess'):
+        if nproc is not None and nproc > 1 and not externals.exists('pprocess'):
             raise RuntimeError("The 'pprocess' module is required for "
                                "multiprocess searchlights. Please either "
                                "install python-pprocess, or reduce `nproc` "
@@ -164,7 +173,10 @@ class Searchlight(BaseSearchlight):
     interest, which is ran at each spatial location.
     """
 
-    def __init__(self, datameasure, queryengine, add_center_fa=False, **kwargs):
+    def __init__(self, datameasure, queryengine, add_center_fa=False,
+                 results_backend='native',
+                 tmp_prefix='tmpsl',
+                 **kwargs):
         """
         Parameters
         ----------
@@ -177,12 +189,26 @@ class Searchlight(BaseSearchlight):
           seed (e.g. sphere center) for the respective ROI. If True, the
           attribute is named 'roi_seed', the provided string is used as the name
           otherwise.
+        results_backend : ('native', 'hdf5'), optional
+          Specifies the way results are provided back from a processing block
+          in case of nproc > 1. 'native' is pickling/unpickling of results by
+          pprocess, while 'hdf5' would use h5save/h5load functionality.
+          'hdf5' might be more time and memory efficient in some cases.
+        tmp_prefix : str, optional
+          If specified -- serves as a prefix for temporary files storage
+          if results_backend == 'hdf5'.  Thus can specify the directory to use
+          (trailing file path separator is not added automagically).
         **kwargs
           In addition this class supports all keyword arguments of its
           base-class :class:`~mvpa2.measures.searchlight.BaseSearchlight`.
         """
         BaseSearchlight.__init__(self, queryengine, **kwargs)
         self.datameasure = datameasure
+        self.results_backend = results_backend.lower()
+        if self.results_backend == 'hdf5':
+            # Assure having hdf5
+            externals.exists('h5py', raise_=True)
+        self.tmp_prefix = tmp_prefix
         if isinstance(add_center_fa, str):
             self.__add_center_fa = add_center_fa
         elif add_center_fa:
@@ -195,14 +221,16 @@ class Searchlight(BaseSearchlight):
             prefixes=prefixes
             + _repr_attrs(self, ['datameasure'])
             + _repr_attrs(self, ['add_center_fa'], default=False)
+            + _repr_attrs(self, ['results_backend'], default='native')
             )
 
 
     def _sl_call(self, dataset, roi_ids, nproc):
         """Classical generic searchlight implementation
         """
+        assert(self.results_backend in ('native', 'hdf5'))
         # compute
-        if nproc > 1:
+        if nproc is not None and nproc > 1:
             # split all target ROIs centers into `nproc` equally sized blocks
             nproc_needed = min(len(roi_ids), nproc)
             roi_blocks = np.array_split(roi_ids, nproc_needed)
@@ -217,10 +245,11 @@ class Searchlight(BaseSearchlight):
                       % nproc_needed)
             compute = p_results.manage(
                         pprocess.MakeParallel(self._proc_block))
-            for block in roi_blocks:
+            for iblock, block in enumerate(roi_blocks):
                 # should we maybe deepcopy the measure to have a unique and
                 # independent one per process?
-                compute(block, dataset, copy.copy(self.__datameasure))
+                compute(block, dataset, copy.copy(self.__datameasure),
+                        iblock=iblock)
 
             # collect results
             results = []
@@ -230,13 +259,14 @@ class Searchlight(BaseSearchlight):
                 roi_sizes = None
 
             for r, rsizes in p_results:
-                results += r
+                results += self.__handle_results(r)
                 if not roi_sizes is None:
                     roi_sizes += rsizes
         else:
             # otherwise collect the results in a list
             results, roi_sizes = \
                     self._proc_block(roi_ids, dataset, self.__datameasure)
+            results = self.__handle_results(results)
 
         if __debug__ and 'SLC' in debug.active:
             debug('SLC', '')            # just newline
@@ -247,17 +277,26 @@ class Searchlight(BaseSearchlight):
         # but be careful: this call also serves as conversion from parallel maps
         # to regular lists!
         # this uses the Dataset-hstack
-        results = hstack(results)
+        result_ds = hstack(results)
+        if self.ca.is_enabled('roi_feature_ids'):
+            self.ca.roi_feature_ids = [r.a.roi_feature_ids for r in results]
 
         if __debug__:
-            debug('SLC', " hstacked shape %s" % (results.shape,))
+            debug('SLC', " hstacked shape %s" % (result_ds.shape,))
 
-        return results, roi_sizes
+        return result_ds, roi_sizes
 
 
-    def _proc_block(self, block, ds, measure):
+    def _proc_block(self, block, ds, measure, iblock='main'):
         """Little helper to capture the parts of the computation that can be
         parallelized
+
+        Parameters
+        ----------
+        iblock
+          Critical for generating non-colliding temp filenames in case
+          of hdf5 backend.  Otherwise RNGs of different processes might
+          collide in their temporary file names leading to problems.
         """
         if __debug__:
             debug_slc_ = 'SLC_' in debug.active
@@ -288,7 +327,14 @@ class Searchlight(BaseSearchlight):
                 roi.fa[self.__add_center_fa] = roi_seed
 
             # compute the datameasure and store in results
-            results.append(measure(roi))
+            res = measure(roi)
+            if self.ca.is_enabled('roi_feature_ids'):
+                if not is_datasetlike(res):
+                    res = Dataset(np.atleast_1d(res))
+                # add roi feature ids to intermediate result dataset for later
+                # aggregation
+                res.a['roi_feature_ids'] = roi_fids
+            results.append(res)
 
             # store the size of the roi dataset
             if not roi_sizes is None:
@@ -301,6 +347,20 @@ class Searchlight(BaseSearchlight):
                        roi.nfeatures,
                        float(i+1)/len(block)*100,), cr=True)
 
+        if self.results_backend == 'native':
+            pass                        # nothing special
+        elif self.results_backend == 'hdf5':
+            # store results in a temporary file and return a filename
+            results_file = tempfile.mktemp(prefix=self.tmp_prefix,
+                                           suffix='-%s.hdf5' % iblock)
+            if __debug__:
+                debug('SLC', "Storing results into %s" % results_file)
+            h5save(results_file, results)
+            if __debug__:
+                debug('SLC_', "Results stored")
+            results = results_file
+        else:
+            raise RuntimeError("Must not reach this point")
         return results, roi_sizes
 
 
@@ -308,6 +368,21 @@ class Searchlight(BaseSearchlight):
         """Set the datameasure"""
         self.untrain()
         self.__datameasure = datameasure
+
+    def __handle_results(self, results):
+        if self.results_backend == 'hdf5':
+            # 'results' must be just a filename
+            assert(isinstance(results, str))
+            if __debug__:
+                debug('SLC', "Loading results from %s" % results)
+            results_data = h5load(results)
+            os.unlink(results)
+            if __debug__:
+                debug('SLC_', "Loaded results of len=%d from"
+                      % len(results_data))
+            return results_data
+        else:
+            return results
 
     datameasure = property(fget=lambda self: self.__datameasure,
                            fset=__set_datameasure)
